requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.7.0=pypi_0
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_9
brotli-bin=1.0.9=h5eee18b_9
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
cfitsio=3.470=h5893167_7
chardet=3.0.4=py39h06a4308_1003
charls=2.2.0=h2531618_0
charset-normalizer=3.4.1=pypi_0
click=8.1.7=py39h06a4308_0
cloudpickle=3.0.0=py39h06a4308_0
contourpy=1.3.0=pypi_0
cycler=0.10.0=py39h06a4308_0
cytoolz=1.0.1=py39h5eee18b_0
dask-core=2024.5.0=py39h06a4308_0
dav1d=1.2.1=h5eee18b_0
decorator=4.4.0=py_0
docker-pycreds=0.4.0=pypi_0
en-core-web-sm=2.1.0=pypi_0
eval-type-backport=0.2.2=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.56.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.12.0=py39h06a4308_0
giflib=5.2.2=h5eee18b_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
h5py=3.12.1=py39h5842655_1
hdf5=1.14.5=h2b7332f_2
idna=3.10=pypi_0
imagecodecs=2024.9.22=py39h73404f3_0
imageio=2.37.0=py39h06a4308_0
importlib-metadata=8.5.0=py39h06a4308_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=pypi_0
joblib=1.4.2=py39h06a4308_0
jpeg=9e=h5eee18b_3
jxrlib=1.1=h7b6447c_2
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libaec=1.1.3=h6a678d5_0
libavif=1.1.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcurl=8.12.1=hc9e6f67_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=hffd6297_1
libwebp-base=1.3.2=h5eee18b_1
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
meson=1.7.0=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
mpmath=1.1.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
networkx=2.3=py_0
nose=1.3.7=pyhd3eb1b0_1008
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.16=h5eee18b_0
packaging=24.2=py39h06a4308_0
partd=1.4.2=py39h06a4308_0
pillow=11.1.0=py39hcea889d_0
pip=25.0=py39h06a4308_0
plac=0.9.6=pyhd3eb1b0_1
platformdirs=4.3.7=pypi_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=pypi_0
pybind11-abi=4=hd3eb1b0_1
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pyparsing=2.4.2=py_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2019.2=py_0
pywavelets=1.5.0=py39ha9d4c09_0
pyyaml=6.0.2=py39h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-image=0.19.3=py39h6a678d5_1
scikit-learn=1.6.1=py39h6a678d5_0
scipy=1.13.1=py39h5f9d8c6_1
sentry-sdk=2.25.1=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.8.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.2=pypi_0
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=3.5.0=py39h2f386ee_0
tifffile=2023.4.12=py39h06a4308_0
tk=8.6.14=h39e8969_0
toolz=1.0.0=py39h06a4308_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
triton=3.2.0=pypi_0
typing-extensions=4.13.0=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025a=h04d1e81_0
urllib3=2.4.0=pypi_0
wandb=0.19.9=pypi_0
wheel=0.45.1=py39h06a4308_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zfp=1.0.0=h6a678d5_0
zipp=3.21.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_1
zlib-ng=2.0.7=h5eee18b_0
zstd=1.5.6=hc292b87_0
